// src/api/__generated__/wallet/model/userSearchItemResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 用户搜索结果项
 */
export interface UserSearchItemResponse {
  /** 用户ID */
  userId?: string;
  /** 用户昵称 */
  nickname?: string;
  /** 用户头像URL */
  avatarUrl?: string;
  /** 用户简介 */
  profile?: string;
  /** 短剧数量 */
  shortDramaCount?: string;
  /** 作品数量 */
  workCount?: string;
  /** 粉丝数量 */
  followerCount?: string;
  /** 获赞数量 */
  totalLikeCount?: string;
  /** 当前登录用户与搜索结果用户的关注关系：SELF-本人 MUTUAL-双方互相关注 FOLLOWING-我关注了对方，对方未关注我 FOLLOW_BACK-对方关注了我，我未关注对方 NONE-双方均未关注 */
  followStatus?: string;
}