// src/api/__generated__/wallet/userwallet-user/userwallet-user.ts
export type searchUsersResponse200 = {
  data: ResultPageResponseUserSearchItemResponse;
  status: 200;
};

export type searchUsersResponse400 = {
  data: ResultObject;
  status: 400;
};

export type searchUsersResponse405 = {
  data: ResultObject;
  status: 405;
};

export type searchUsersResponse500 = {
  data: ResultObject;
  status: 500;
};

export type searchUsersResponseSuccess = searchUsersResponse200 & {
  headers: Headers;
};
export type searchUsersResponseError = (
  | searchUsersResponse400
  | searchUsersResponse405
  | searchUsersResponse500
) & {
  headers: Headers;
};

export type searchUsersResponse =
  | searchUsersResponseSuccess
  | searchUsersResponseError;